% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_tasks.R
\name{active_tasks}
\alias{active_tasks}
\title{active tasks}
\usage{
active_tasks(cushion, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to one of the HTTP verbs (e.g,.
\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{PUT}},
etc.)}
}
\value{
JSON as a character string or a list (determined by the
\code{as} parameter)
}
\description{
active tasks
}
\examples{
\dontrun{
# Create a CouchDB connection client
(x <- Cushion$new())

active_tasks(x)
active_tasks(x, as = 'json')
}
}

