% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_nsofa_drug_start_end_times.R
\name{align_nsofa_drug_start_end_times}
\alias{align_nsofa_drug_start_end_times}
\title{Aligns the start and end times for drun dosages}
\usage{
align_nsofa_drug_start_end_times(df)
}
\arguments{
\item{df}{a dataste containing drug dosages}
}
\value{
A data frame indicating if a subjects was on a given drug during a given hour
}
\description{
Ensures that the end time dosage occurs after the start time dosage and converts the dosage time to binary
1 = yes, 0 = no to indicate if the subject was on that drug during a q1hr

\itemize{
\item Ensures that the end time dosage occurs after the start time dosage
\item Converts the dosage time to 1 = yes, 0 = no to indicate if the subject was on that drug during a q1hr
\item Rounds the dosage time down to the nearest hour
}
}
