% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldocs.r
\name{alldocs}
\alias{alldocs}
\title{List all docs in a given database.}
\usage{
alldocs(cushion, dbname, asdf = TRUE, descending = NULL, startkey = NULL,
  endkey = NULL, limit = NULL, include_docs = NULL, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{Database name. (charcter)}

\item{asdf}{Return as data.frame? defaults to TRUE (logical)}

\item{descending}{Return in descending order? (logical)}

\item{startkey}{Document ID to start at. (character)}

\item{endkey}{Document ID to end at. (character)}

\item{limit}{Number document IDs to return. (numeric)}

\item{include_docs}{If TRUE, returns docs themselves, in addition to IDs (logical)}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
List all docs in a given database.
}
\examples{
\dontrun{
(x <- Cushion$new())

alldocs(x, dbname="sofadb")
alldocs(x, dbname="sofadb", as='json')
alldocs(x, dbname="mydb", limit=2)
alldocs(x, dbname="mydb", limit=2, include_docs="true")
library('httr')
alldocs(x, dbname="sofadb", config=verbose())
}
}

