% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.r
\name{attach_create}
\alias{attach_create}
\title{Include an attachment either on an existing or new document}
\usage{
attach_create(cushion, dbname, docid, attachment, attname, ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{(charcter) Database name. Required.}

\item{docid}{(charcter) Document ID. Required.}

\item{attachment}{(charcter) The attachment object name. Required.}

\item{attname}{(charcter) Attachment name. Required.}

\item{...}{Curl args passed on to \code{\link[httr]{POST}}}
}
\value{
a list
}
\description{
Include an attachment either on an existing or new document
}
\examples{
\dontrun{
(x <- Cushion$new())

if ("drinksdb" \%in\% db_list(x)) {
  invisible(db_delete(x, dbname="drinksdb"))
}
db_create(x, dbname='drinksdb')

# put on to an existing document
doc <- '{"name":"stuff", "drink":"soda"}'
doc_create(x, dbname="drinksdb", doc=doc, docid="asoda")
myattachment <- "just a simple text string"
myattachment <- mtcars
attach_create(x, dbname="drinksdb", docid="asoda",
  attachment=myattachment, attname="mtcarstable.csv")
}
}

