% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getattach.r
\name{attach_get}
\alias{attach_get}
\title{Get an attachment}
\usage{
attach_get(...)
}
\arguments{
\item{...}{ignored}
}
\description{
This function is defunct. See \code{\link[=doc_attach_get]{doc_attach_get()}}
}
\keyword{internal}
