% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.r
\name{authentication}
\alias{authentication}
\alias{cushion}
\alias{cushions}
\title{Put a cushion on the sofa, and get cushion info}
\usage{
cushion(name, user = NULL, pwd = NULL, base = NULL, type = NULL,
  port = 5984)

cushions(authfile = "~/.sofa-auth")
}
\arguments{
\item{name}{Name for the cushion. This is what you'll call in sofa functions to get these
details.}

\item{user}{A user name}

\item{pwd}{A password}

\item{base}{A base URL, not needed if \code{type=localhost|cloudant|iriscouch}. Though
you can pass a base URL in here to override anything done internally.}

\item{type}{One of localhost, cloudant, iriscouch, or \code{NULL}. This is what's
used to determine how to structure the URL to make the request. If left to the
default of \code{NULL}, you must pass in a base URL.}

\item{port}{Port. Only applies when type is localhost. Default: 5984}

\item{authfile}{Path to file with cushions (authentication details)}
}
\description{
That is, set up config for remote CouchDB databases, or get auth info
}
\details{
Setup for authentication:
For localhost you don't need to authenticate, but of course you may have set
up a username and password in which case use 'localhost'.

Others supported are 'cloudant' and 'iriscouch'.

You can use other named username/password sets too.

You can permanently store your auth details in a hidden file ~/.sofa-auth by putting in
entries like this, each on a new line:

list(name="iriscouch", user="jane_doe", pwd="my_password", type="iriscouch")

Though if you don't want to store them permanently, you can use the \code{cushion}
function instead. See examples below on how to do this. Using \code{cushion} only
stores them for the current session.

\code{cushions()} Looks first in the local environment SofaAuthCache, and if finds nothing,
looks in your \code{~/.sofa-auth} file, or wherver you set \code{authfilie} parameter to.

Beware that you should use unique names for each cushion, that is, in the \code{name} parameter.
}
\examples{
\dontrun{
cushion('cloudant', user='name', pwd='pwd', type="cloudant")
cushion('iriscouch', user='name', pwd='pwd', type="iriscouch")
cushion('julies_iris', user='name', pwd='pwd', type="iriscouch")
cushion('adfafafafadf', user='name', pwd='pwd', type="localhost", port=2300)
cushions()

cushion("oceancouch", base="http://104.236.176.205", port=5984)
}
}

