% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bulk_update.R
\name{bulk_update}
\alias{bulk_update}
\title{Create documents via the bulk API}
\usage{
bulk_update(doc, cushion = "localhost", dbname, docid = NULL,
  how = "rows", as = "list", ...)
}
\arguments{
\item{doc}{For now, a data.frame only}

\item{cushion}{A cushion name}

\item{dbname}{Database name3
\code{\link{doc_get}}, the XML is given back and you can parse it as normal.}

\item{docid}{Document IDs, ignored for now, eventually, you can pass in a list, or
vector to be the ids for each document created. Has to be the same length as the
number of documents.}

\item{how}{(character) One of rows (default) or columns. If rows, each row becomes a
separate document; if columns, each column becomes a separate document.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Create documents via the bulk API
}
\examples{
\dontrun{
row.names(mtcars) <- NULL
db_delete(dbname="bulktest")
db_create(dbname="bulktest")
bulk_create(mtcars, dbname="bulktest")

# modify mtcars
mtcars$letter <- sample(letters, NROW(mtcars), replace = TRUE)
bulk_update(mtcars, dbname="bulktest")

# change again
mtcars$num <- 89
bulk_update(mtcars, dbname="bulktest")
}
}

