% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.r
\name{changes}
\alias{changes}
\title{List changes to a database.}
\usage{
changes(cushion, dbname, descending = NULL, startkey = NULL,
  endkey = NULL, since = NULL, limit = NULL, include_docs = NULL,
  feed = "normal", heartbeat = NULL, filter = NULL, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{Database name. (charcter)}

\item{descending}{Return in descending order? (logical)}

\item{startkey}{Document ID to start at. (character)}

\item{endkey}{Document ID to end at. (character)}

\item{since}{Start the results from the change immediately after the given sequence number.}

\item{limit}{Number document IDs to return. (numeric)}

\item{include_docs}{If "true", returns docs themselves, in addition to IDs (character)}

\item{feed}{Select the type of feed. One of normal, longpoll, or continuous. See description. (character)}

\item{heartbeat}{Period in milliseconds after which an empty line is sent in the results.
Only applicable for longpoll or continuous feeds. Overrides any timeout to keep the
feed alive indefinitely. (numeric (milliseconds))}

\item{filter}{Reference a filter function from a design document to selectively get updates.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Of course it doesn't make much sense to use certain options in
   _changes. For example, using feed=longpoll or continuous doesn't make much sense
   within R itself.
}
\examples{
\dontrun{
(x <- Cushion$new())

changes(x, dbname="sofadb")
changes(x, dbname="sofadb", as='json')
changes(x, dbname="sofadb", limit=2)
}
}

