% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createdb}
\alias{createdb}
\title{Create a database.}
\usage{
createdb(cushion = "localhost", dbname, delifexists = FALSE, ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name}

\item{delifexists}{If TRUE, delete any database of the same name before creating it.
This is useful for testing. Default is FALSE.}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Create a database.
}
\examples{
\donttest{
createdb(dbname='leothelion')
listdbs() # see if its there now

## or setting username and password in cushion() call
createdb("cloudant", "mustache")

## iriscouch
createdb("iriscouch", "beard")
}
}

