% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cushion.R
\docType{data}
\name{Cushion}
\alias{Cushion}
\title{sofa connection client}
\arguments{
\item{name}{(character) Name for the cushion. This is what you'll call in sofa functions to get these
details.}

\item{host}{(character) A base URL, not needed if \code{type=localhost|cloudant|iriscouch}. Though
you can pass a base URL in here to override anything done internally.}

\item{port}{(numeric) Port. Only applies when type is localhost. Default: 5984}

\item{path}{(character) context path that is appended to the end of the url.
Default: NULL, ignored}

\item{user}{(character) A user name}

\item{pwd}{(character) A password}

\item{type}{(character) One of localhost, cloudant, iriscouch, or \code{NULL}. This is what's
used to determine how to structure the URL to make the request. If left to the
default of \code{NULL}, you must pass in a base URL.}
}
\description{
sofa connection client
}
\details{
\strong{Methods}
  \describe{
    \item{\code{ping()}}{
      Ping the CouchDB server
    }
    \item{\code{make_url()}}{
      Construct full base URL from the pieces in the connection object
    }
  }
}
\examples{
\dontrun{
# Create a CouchDB connection client
(x <- Cushion$new(name = "foobar"))

## metadata
x$host
x$path
x$name
x$port
x$type

## ping the CouchDB server
x$ping()

## CouchDB server statistics
stats(x)

## database info
db_info(x, "bulktest")

## list dbs
db_list(x)

## all docs
alldocs(x, "bulktest", limit = 3)

## changes
changes(x, "bulktest")

# create database
db_create(x, "stuff")

# add documents to a database
db_create(x, "sofadb")
doc1 <- '{"name": "drink", "beer": "IPA", "score": 5}'
doc_create(x, dbname="sofadb", docid="abeer", doc1)

# bulk create
db_create(x, "mymtcars")
bulk_create(x, dbname="mymtcars", doc = mtcars)
db_list(x)

# Using Cloudant
# (z <- Cushion$new('foobar', type="cloudant",
#   user = 'ropensci', pwd = 'agXX9ypzYDD3FqcLKvWGQiFZx'))
# z$db_info()
# z$alldocs(limit = 2)

## create a cloudant db
# (z <- Cushion$new('stuff', type="cloudant",
#     user = 'ropensci', pwd = 'agXX9ypzYDD3FqcLKvWGQiFZx'))
# z$db_create()
}
}
\keyword{datasets}

