% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.r
\name{databases}
\alias{databases}
\title{Work with databases in CouchDB.}
\description{
Work with databases in CouchDB.
}
\details{
There are the following functions for working with databases:
\itemize{
\item \code{\link[=db_create]{db_create()}} - Create a database
\item \code{\link[=db_delete]{db_delete()}} - Delete a database
\item \code{\link[=db_info]{db_info()}} - Get info for a database
\item \code{\link[=db_list]{db_list()}} - List databases
\item \code{\link[=db_replicate]{db_replicate()}} - Replicate a database from one couch to another
}
}
