% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list.r
\name{db_list}
\alias{db_list}
\title{List all databases.}
\usage{
db_list(cushion, simplify = TRUE, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{simplify}{(logical) Simplify to character vector, ignored
if \code{as="json"}}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to one of the HTTP verbs (e.g,.
\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{PUT}},
etc.)}
}
\value{
JSON as a character string or a list (determined by the
\code{as} parameter)
}
\description{
List all databases.
}
\examples{
\dontrun{
(x <- Cushion$new())

db_list(x)
db_list(x, as = 'json')
}
}
