% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbinfo}
\alias{dbinfo}
\title{List database info.}
\usage{
dbinfo(cushion = "localhost", dbname, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
List database info.
}
\examples{
\donttest{
# local databasees
dbinfo(dbname="sofadb")
dbinfo(dbname="sofadb", as='json')

# a database on cloudant or iriscouch
dbinfo("cloudant", "gaugesdb_ro")
dbinfo("iriscouch", "helloworld")
}
}

