% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deldoc}
\alias{deldoc}
\title{Delete a document in a database.}
\usage{
deldoc(cushion = "localhost", dbname, docid, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a document in a database.
}
\examples{
\donttest{
doc3 <- "<top><a/><b/><c><d/><e>bob</e></c></top>"
writedoc(dbname="sofadb", doc=doc3, docid="newnewxml")
deldoc(dbname="sofadb", docid="newnewxml")
getdoc(dbname="sofadb", docid="newnewxml")

# wrong docid name
writedoc(dbname="sofadb", doc=doc3, docid="newxml")
deldoc(dbname="sofadb", docid="wrongname")

# remote
doc3 <- "<top><a/><b/><c><d/><e>bob</e></c></top>"
writedoc("cloudant", dbname="mustache", doc=doc3, docid="bbbb")
getdoc("cloudant", dbname='mustache', docid='bbbb')
deldoc("cloudant", dbname='mustache', docid='bbbb')
}
}

