% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deletedb}
\alias{deletedb}
\title{Delete a database.}
\usage{
deletedb(dbname, endpoint = "localhost", port = 5984, username = NULL,
  pwd = NULL, ...)
}
\arguments{
\item{dbname}{Database name}

\item{endpoint}{One of localhost, cloudant, or iriscouch. For local work
use the default localhost. For cloudant or iriscouch you will need accounts
with those database services.}

\item{port}{The port to use. Only applicable if using endpoint="localhost".}

\item{username}{Your cloudant or iriscouch username}

\item{pwd}{Your cloudant or iriscouch password}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a database.
}
\examples{
\donttest{
# local databasees
## create database first, then delete
createdb('newdb')
deletedb('newdb')

## with curl info while doing request
library('httr')
createdb('newdb')
deletedb('newdb', config=verbose())
}
}

