% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc_bulk.R
\name{doc_bulk}
\alias{doc_bulk}
\title{Create documents via the bulk API}
\usage{
doc_bulk(doc, cushion = "localhost", dbname, docid = NULL, how = "rows",
  as = "list", ...)
}
\arguments{
\item{doc}{For now, a data.frame only}

\item{cushion}{A cushion name}

\item{dbname}{Database name3
\code{\link{doc_get}}, the XML is given back and you can parse it as normal.}

\item{docid}{Document IDs, ignored for now, eventually, you can pass in a list, or
vector to be the ids for each document created. Has to be the same length as the
number of documents.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Create documents via the bulk API
}
\examples{
\dontrun{
db_delete(dbname="bulktest")
db_create(dbname="bulktest")
doc_bulk(mtcars, dbname="bulktest")

db_create(dbname="bulktest2")
doc_bulk(iris, dbname="bulktest2")
}
}

