% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc_delete.r
\name{doc_delete}
\alias{doc_delete}
\title{Delete a document in a database.}
\usage{
doc_delete(cushion = "localhost", dbname, docid, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a document in a database.
}
\examples{
\dontrun{
doc3 <- "<top><a/><b/><c><d/><e>bob</e></c></top>"
doc_create(dbname="sofadb", doc=doc3, docid="newnewxml")
doc_delete(dbname="sofadb", docid="newnewxml")
doc_delete(dbname="sofadb", docid="newnewxml")

# wrong docid name
doc_create(dbname="sofadb", doc=doc3, docid="newxml")
doc_delete(dbname="sofadb", docid="wrongname")

# remote
doc3 <- "<top><a/><b/><c><d/><e>bob</e></c></top>"
doc_create("cloudant", dbname="mustache", doc=doc3, docid="bbbb")
doc_get("cloudant", dbname='mustache', docid='bbbb')
doc_delete("cloudant", dbname='mustache', docid='bbbb')

# On arbitrary remote server
doc <- list(name = "drink", beer = "IPA")
doc_create("oceancouch", dbname="beard", doc=doc, docid="beer")
doc_delete("oceancouch", dbname="beard", docid="beer")
}
}

