% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doc_update}
\alias{doc_update}
\title{Update a document.}
\usage{
doc_update(cushion = "localhost", dbname, doc, docid, rev, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name3}

\item{doc}{Document content}

\item{docid}{Document ID}

\item{rev}{Revision id.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Update a document.
}
\details{
Internally, this function adds in the docid and revision id, required to do a
document update.
}
\examples{
\donttest{
doc1 <- '{"name":"drink","beer":"IPA"}'
doc_create(dbname="sofadb", doc=doc1, docid="a_beer")
doc_get(dbname = "sofadb", docid = "a_beer")
revs <- revisions(dbname = "sofadb", docid = "a_beer")
doc2 <- '{"name":"drink","beer":"IPA","note":"yummy","note2":"yay"}'
doc_update(dbname="sofadb", doc=doc2, docid="a_beer", rev=revs[1])
revisions(dbname = "sofadb", docid = "a_beer")
}
}

