\name{dude_ping}
\alias{dude_ping}
\title{Ping the couchdb server.}
\usage{
  dude_ping(endpoint = "http://127.0.0.1", port = 5984)
}
\arguments{
  \item{endpoint}{the endpoint, defaults to localhost
  (http://127.0.0.1)}

  \item{port}{port to connect to, defaults to 5984}
}
\description{
  Ping the couchdb server.
}
\examples{
dude_ping()
}

