% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_max_nsofa_score_within_n_hours_of_admission.R
\name{get_max_nosfa_score_within_n_hours_of_admission}
\alias{get_max_nosfa_score_within_n_hours_of_admission}
\title{Calculate maximum nsofa score within a time range}
\usage{
get_max_nosfa_score_within_n_hours_of_admission(
  nsofa_scores,
  min_hour,
  max_hour
)
}
\arguments{
\item{nsofa_scores}{a dataframe contianing q1hr nsofa scores}

\item{min_hour}{lower timepoint}

\item{max_hour}{upper timepoint}
}
\value{
A data frame with the max nosfa score within a given time range
}
\description{
Calculate maximum nsofa score within a time range
}
