% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsofa_inotropes.R
\name{get_nsofa_inotropes}
\alias{get_nsofa_inotropes}
\title{Calculate Inotropes score}
\usage{
get_nsofa_inotropes(read_medications, child_encounter)
}
\arguments{
\item{read_medications}{the raw child medication csv file obtained from the IDR}

\item{child_encounter}{a dataframe returned by  \code{\link{get_nsofa_child_encounter}}}
}
\value{
A data frame with the calculated inotropes score
}
\description{
Calculate Inotropes score
}
