% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsofa_oxygenation.R
\name{get_nsofa_oxygenation}
\alias{get_nsofa_oxygenation}
\title{Calculates oxygenation score}
\usage{
get_nsofa_oxygenation(read_flowsheets, respiratory_devices, child_encounter)
}
\arguments{
\item{read_flowsheets}{the raw flowsheets csv file obtained from the IDR}

\item{respiratory_devices}{an excel file containing the categorized respiratory devices}

\item{child_encounter}{a df returned by \code{\link{get_nsofa_child_encounter}}}
}
\value{
A dataframe containing the oxygenation score
}
\description{
Sets the respiratory status based on the oxygenation or vent settings device present
}
