% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsofa_platelets.R
\name{get_nsofa_platelets}
\alias{get_nsofa_platelets}
\title{Calculate platelets score}
\usage{
get_nsofa_platelets(read_child_labs, child_encounter)
}
\arguments{
\item{read_child_labs}{the raw child labs csv file obtained from the IDR}

\item{child_encounter}{a dataframe returned by  \code{\link{get_nsofa_child_encounter}}}
}
\value{
A data frame with the calculated platelets score
}
\description{
Ingests the raw child labs and calculates the coagulation score based on the platelet count.

\itemize{
\item Uses the last recorded inferred_specimen_datetime within an hour
}
}
