% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsofa_steroids.R
\name{get_nsofa_steroids}
\alias{get_nsofa_steroids}
\title{Calculate steroid score}
\usage{
get_nsofa_steroids(read_child_medications, child_encounter)
}
\arguments{
\item{read_child_medications}{the raw child medication csv file obtained from the IDR}

\item{child_encounter}{a dataframe returned by  \code{\link{get_nsofa_child_encounter}}}
}
\value{
A data frame with the calculated steroid score
}
\description{
Calculate steroid score
}
