% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psofa_cardiovascular.R
\name{get_psofa_cv_by_vasoactive_infusion}
\alias{get_psofa_cv_by_vasoactive_infusion}
\title{Calculate cardiovascular score based on vasoactive infusions}
\usage{
get_psofa_cv_by_vasoactive_infusion(read_medications, expanded_child_encounter)
}
\arguments{
\item{read_medications}{the raw medications data obtained from the IDR}

\item{expanded_child_encounter}{a df returned by \code{\link{get_psofa_child_encounter}}}
}
\value{
A df containing the calculated cardiovascular score based on vasoactive infusions
}
\description{
Calculate cardiovascular score based on vasoactive infusions
}
