% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psofa_hepatic.R
\name{get_psofa_hepatic}
\alias{get_psofa_hepatic}
\title{Calculate hepatic score}
\usage{
get_psofa_hepatic(read_child_labs)
}
\arguments{
\item{read_child_labs}{the raw child labs csv file obtained from the IDR}
}
\value{
A data frame with the calculated hepatic score
}
\description{
Ingests the raw child labs and calculates the hepatic score based on the bilirubin level.

\itemize{
\item Use the last recorded inferred_specimen_datetime within an hour
\item If there are multiple values for the last recorded inferred_specimen_datetime then use the max hepatic_score
}
}
