% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psofa_respiratory_support.R
\name{get_psofa_respiratory}
\alias{get_psofa_respiratory}
\title{Calculate respiratory score}
\usage{
get_psofa_respiratory(
  expanded_child_encounter,
  pao2_fio2_resp_score,
  complete_fio2_value,
  spo2_value,
  respiratory_support_status,
  nasal_flow_rate_state
)
}
\arguments{
\item{expanded_child_encounter}{a df returned by \code{\link{get_psofa_child_encounter}}}

\item{pao2_fio2_resp_score}{a df returned by \code{\link{get_psofa_pao2_fio2_resp_score}}}

\item{complete_fio2_value}{a df returned by \code{\link{get_psofa_fio2_value}}}

\item{spo2_value}{a df returned by \code{\link{get_psofa_spo2_value}}}

\item{respiratory_support_status}{a df returned by \code{\link{get_psofa_respiratory_status}}}

\item{nasal_flow_rate_state}{a df returned by \code{\link{get_psofa_nasal_state}}}
}
\value{
A dataframe containing the respiratory scores
}
\description{
Calculate respiratory score
}
