% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psofa_respiratory_support.R
\name{get_psofa_spo2_value}
\alias{get_psofa_spo2_value}
\title{Extract SpO2 Value}
\usage{
get_psofa_spo2_value(read_flowsheets)
}
\arguments{
\item{read_flowsheets}{the raw flowsheet csv file obtained from the IDR}
}
\value{
A dataframe containing the SpO2 values
}
\description{
Extracts the SpO2 value from the vitals flowsheets
\itemize{
\item Priority: SpO2, SpO2 #3, then SpO2 #2

\item when there are multiple values for any given SpO2 within an hour
choose the lowest value
}
}
