% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getattach}
\alias{getattach}
\title{Get an attachment.}
\usage{
getattach(cushion = "localhost", dbname, docid, attname = NULL, ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}

\item{attname}{Attachment name.}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get an attachment.
}
\examples{
\donttest{
getattach(dbname="sofadb", docid="abeer")
}
}

