% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listdbs}
\alias{listdbs}
\title{List all databases.}
\usage{
listdbs(cushion = "localhost", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
List all databases.
}
\examples{
\donttest{
# local databasees
listdbs()
listdbs("cloudant")
listdbs("iriscouch")
}
}

