% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.r
\name{ping}
\alias{ping}
\title{Ping a couchdb server}
\usage{
ping(cushion, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Ping a couchdb server
}
\examples{
\dontrun{
# initialize a couchdb connection
(x <- Cushion$new())

ping(x)
ping(x, as = "json")
}
}

