% Generated by roxygen2 (4.0.2): do not edit by hand
\name{revisions}
\alias{revisions}
\title{Get document revisions.}
\usage{
revisions(cushion = "localhost", dbname, docid, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name}

\item{docid}{Document ID}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get document revisions.
}
\examples{
\donttest{
revisions(dbname="sofadb", docid="a_beer")
revisions(dbname="sofadb", docid="a_beer", as='json')
}
}

