% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\title{session}
\usage{
session(cushion, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
session
}
\examples{
\dontrun{
# Create a CouchDB connection client
(x <- Cushion$new())

session(x)
session(x, as = 'json')
}
}

