% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sf_attach}
\alias{sf_attach}
\title{Include an attachment either on an existing or new document.}
\usage{
sf_attach(endpoint = "http://127.0.0.1", port = 5984, dbname, docid,
  attachment, attname)
}
\arguments{
\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}

\item{attachment}{The attachment object name}

\item{attname}{Attachment name.}
}
\description{
Include an attachment either on an existing or new document.
}
\examples{
\donttest{
# put on to an existing document
doc <- '{"name":"guy","beer":"anybeerisfine"}'
writedoc(dbname="sofadb", doc=doc, docid="guysbeer")
myattachment <- "just a simple text string"
myattachment <- mtcars
sf_attach(dbname="sofadb", docid="guysbeer", attachment=myattachment, attname="mtcarstable.csv")
}
}

