% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sf_head}
\alias{sf_head}
\title{Get header info for a document.}
\usage{
sf_head(endpoint = "localhost", port = 5984, dbname, docid,
  username = NULL, pwd = NULL, ...)
}
\arguments{
\item{port}{port to connect to, defaults to 5984}

\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}

\item{username}{Your cloudant or iriscouch username}

\item{pwd}{Your cloudant or iriscouch password}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get header info for a document.
}
\examples{
\donttest{
sf_head(dbname="sofadb", docid="beer")
}
}

