% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sofa_deletedb}
\alias{sofa_deletedb}
\title{Delete a database.}
\usage{
sofa_deletedb(dbname, endpoint = "localhost", port = 5984,
  username = NULL, pwd = NULL, ...)
}
\arguments{
\item{endpoint}{One of localhost, cloudant, or iriscouch. For local work
use the default localhost. For cloudant or iriscouch you will need accounts
with those database services.}

\item{port}{The port to use. Only applicable if using endpoint="localhost".}

\item{username}{Your cloudant or iriscouch username}

\item{pwd}{Your cloudant or iriscouch password}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a database.
}
\examples{
\donttest{
# local databasees
## create database first, then delete
sofa_createdb('newdb')
sofa_deletedb('newdb')

## with curl info while doing request
library('httr')
sofa_createdb('newdb')
sofa_deletedb('newdb', config=verbose())
}
}

