% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sofa_getdoc}
\alias{sofa_getdoc}
\title{Get a document from a database.}
\usage{
sofa_getdoc(endpoint = "http://127.0.0.1", port = 5984, dbname, docid)
}
\arguments{
\item{endpoint}{the endpoint, defaults to localhost (http://127.0.0.1)}

\item{port}{port to connect to, defaults to 5984}

\item{dbname}{Database name}

\item{docid}{Document ID}
}
\description{
Get a document from a database.
}
\examples{
sofa_getdoc(dbname="twitter_db", docid="9f6950f1ee18ed0f0a2c529c30003ab0")
}

