% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Get stats on your Couchdb instance}
\usage{
stats(cushion, section = NULL, stat = NULL, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{section}{(character) One of couchdb, httpd_request_methods, or
httpd_status_codes. Default is none of them, which returns all sections.}

\item{stat}{(character) Statistic to return. There are many options.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get stats on your Couchdb instance
}
\details{
Note that if you provide \code{section} you have to provide
\code{stat}, or the function drops those variables, and defaults to
all statistics.

Note that this isn't working with CouchDB v2, not sure why
}
\examples{
\dontrun{
stats(x)
stats(x, as = "json")

# couchdb stats
stats(x, section="couchdb", stat="database_writes")
stats(x, section="couchdb", stat = "open_os_files")

# request methods
stats(x, section="httpd_request_methods", stat="GET")
stats(x, section="httpd_request_methods", stat="POST")

# request status codes
stats(x, section="httpd_status_codes", stat=200)
stats(x, section="httpd_status_codes", stat=400)
}
}

