% Generated by roxygen2 (4.0.2): do not edit by hand
\name{upload}
\alias{upload}
\title{Upload a local database to a remote database server, e.g., Cloudant, Iriscouch}
\usage{
upload(to = "cloudant", port = 5984, dbname, username = NULL,
  pwd = NULL, createdb = FALSE)
}
\arguments{
\item{to}{Remote service name to upload to. One of cloudant, iriscouch.}

\item{port}{port to connect to, defaults to 5984}

\item{dbname}{Database name.}

\item{username}{Your cloudant or iriscouch username}

\item{pwd}{Your cloudant or iriscouch password}

\item{createdb}{If TRUE, the function creates the db on the remote server before
uploading. The db has to exist before uploading, so either you do it separately
or this fxn can do it for you. Default = FALSE}
}
\description{
Upload a local database to a remote database server, e.g., Cloudant, Iriscouch
}
\examples{
\donttest{
# Create a database locally
createdb('hello_earth')

# Upload to a remote server
upload(to="cloudant", dbname="hello_earth", createdb=TRUE)
upload(to="iriscouch", dbname="hello_earth", createdb=TRUE)
}
}

