% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uuids}
\alias{uuids}
\title{Get uuids.}
\usage{
uuids(cushion = "localhost", count = 1, ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{count}{(numeric) Number of uuids to return. Default: 1}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get uuids.
}
\examples{
\donttest{
uuids()
uuids("cloudant")
uuids("iriscouch")
}
}

