% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{views}
\alias{view_create}
\alias{view_create_}
\alias{view_delete}
\alias{view_get}
\alias{view_search}
\alias{views}
\title{Work with design documents (i.e., views) in your CouchDB's.}
\usage{
view_create(cushion, dbname, design_name, fxnname, key = "null",
  value = "doc", as = "list", ...)

view_create_(cushion, dbname, design_name, fxnname, fxn, as = "list", ...)

view_delete(cushion, dbname, design_name, as = "list", ...)

view_get(cushion, dbname, design_name, as = "list", ...)

view_search(cushion, dbname, design_name, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{(character) Database name. required.}

\item{design_name}{(character) Design document name. required.}

\item{fxnname}{(character) A function name. required for \code{view_put}
and \code{view_put_}}

\item{key, value}{(character) a key and value, see Examples and Details}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}

\item{fxn}{(character) a javascript function. required for \code{view_put_}}
}
\description{
Work with design documents (i.e., views) in your CouchDB's.
}
\details{
\code{view_create} is a slightly easier interface to creating
design documents; it just asks for a function name, the key and a
value, then we create the function for you internally. TO have more
flexibility use \code{view_put_} (with underscore on the end) to write the
function yourself.
}
\examples{
\dontrun{
(x <- Cushion$new())

file <- system.file("examples/omdb.json", package = "sofa")
strs <- readLines(file)

## create a database
if ("omdb" \%in\% db_list(x)) {
  invisible(db_delete(x, dbname="omdb"))
}
db_create(x, dbname='omdb')

## add the documents
invisible(db_bulk_create(x, "omdb", strs))

# Create a view, the easy way, but less flexible
view_create(x, dbname='omdb', design_name='view1', fxnname="foobar1")
view_create(x, dbname='omdb', design_name='view2', fxnname="foobar2",
  value="doc.Country")
view_create(x, dbname='omdb', design_name='view5', fxnname="foobar3",
  value="[doc.Country,doc.imdbRating]")

# the harder way, write your own function, but more flexible
view_create_(x, dbname='omdb', design_name='view2',
  fxnname = "stuffthings", fxn = "function(doc){emit(null,doc.Country)}")

# Delete a view
view_delete(x, dbname='omdb', design_name='view1')

# Get info on a design document (i.e., a view)
view_get(x, dbname='omdb', design_name='view2')
view_get(x, dbname='omdb', design_name='view5')

# Search using a view
res <- view_search(x, dbname='omdb', design_name='view2')
head(
  do.call(
    "rbind.data.frame",
    lapply(res$rows, function(x) Filter(length, x))
  )
)

res <- view_search(x, dbname='omdb', design_name='view5')
head(
  structure(do.call(
    "rbind.data.frame",
    lapply(res$rows, function(x) x$value)
  ), .Names = c('Country', 'imdbRating'))
)
}
}

