% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_tasks.R
\name{active_tasks}
\alias{active_tasks}
\title{active tasks}
\usage{
active_tasks(cushion, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{\link{Cushion}} object. Required.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[crul]{HttpClient}}}
}
\value{
JSON as a character string or a list (determined by the
\code{as} parameter)
}
\description{
active tasks
}
\examples{
\dontrun{
# Create a CouchDB connection client
user <- Sys.getenv("COUCHDB_TEST_USER")
pwd <- Sys.getenv("COUCHDB_TEST_PWD")
(x <- Cushion$new(user=user, pwd=pwd))

active_tasks(x)
active_tasks(x, as = 'json')
}
}
