% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldocs.r
\name{alldocs}
\alias{alldocs}
\title{List all docs in a given database.}
\usage{
alldocs(cushion = "localhost", dbname, asdf = TRUE, descending = NULL,
  startkey = NULL, endkey = NULL, limit = NULL, include_docs = NULL,
  as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name. (charcter)}

\item{asdf}{Return as data.frame? defaults to TRUE (logical)}

\item{descending}{Return in descending order? (logical)}

\item{startkey}{Document ID to start at. (character)}

\item{endkey}{Document ID to end at. (character)}

\item{limit}{Number document IDs to return. (numeric)}

\item{include_docs}{If TRUE, returns docs themselves, in addition to IDs (logical)}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
List all docs in a given database.
}
\examples{
\dontrun{
alldocs(dbname="sofadb")
alldocs(dbname="sofadb", as='json')
alldocs(dbname="mydb", limit=2)
alldocs(dbname="mydb", limit=2, include_docs="true")
library('httr')
alldocs(dbname="sofadb", config=verbose())

# different login credentials than normal, just pass in to function call
## you obviously need to fill in some details here, this won't work as is
alldocs("cloudant", dbname='dbname')

# this works for the package author, but not for you
alldocs(cushion="cloudant", dbname='gaugesdb_ro')

# irishcouch
alldocs(cushion="iriscouch", dbname='helloworld')

# any remote couch, this is just a couchc on a DigitalOcean droplet
alldocs(cushion="oceancouch", dbname='mapuris')
}
}

