% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.r
\name{attach_create}
\alias{attach_create}
\title{Include an attachment either on an existing or new document}
\usage{
attach_create(cushion, dbname, docid, attachment, attname, ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{(charcter) Database name. Required.}

\item{docid}{(charcter) Document ID. Required.}

\item{attachment}{(charcter) The attachment object name. Required.}

\item{attname}{(charcter) Attachment name. Required.}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}

\item{as}{(character) One of list (default) or json}
}
\description{
Include an attachment either on an existing or new document
}
\examples{
\dontrun{
(x <- Cushion$new())

# put on to an existing document
doc <- '{"name":"guy","beer":"anybeerisfine"}'
doc_create(x, dbname="sofadb", doc=doc, docid="guysbeer")
myattachment <- "just a simple text string"
myattachment <- mtcars
attach_create(x, dbname="sofadb", docid="guysbeer",
  attachment=myattachment, attname="mtcarstable.csv")
}
}

