% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getattach.r
\name{attach_get}
\alias{attach_get}
\title{Get an attachment.}
\usage{
attach_get(cushion = "localhost", dbname, docid, attname = NULL,
  as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}

\item{attname}{Attachment name.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get an attachment.
}
\examples{
\dontrun{
attach_get(dbname="sofadb", docid="guysbeer")
attach_get("oceancouch", dbname="mapuris", docid="goodbeer")
}
}

