% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authentication}
\alias{authentication}
\alias{cushion}
\alias{profiles}
\title{Put a cushion on the sofa, and get cushion info}
\usage{
cushion(name, user = NULL, pwd = NULL, type, port = NULL)

profiles()
}
\arguments{
\item{name}{Name for the cushion. This is what you'll call in sofa functions to get these
details.}

\item{user}{A user name}

\item{pwd}{A password}

\item{type}{One of localhost, cloudant, or iriscouch. This is what's used to determine
how to structure the URL to make the request.}

\item{port}{Port. Only applies when type is localhost.}
}
\description{
That is, set up config for remote CouchDB databases, or get auth info
}
\details{
Setup for authentication:
For localhost you don't need to authenticate, but of course you may have set
up a username and password in which case use 'localhost'.

Others supported are 'cloudant' and 'iriscouch'.

You can use other named username/password sets too.

You can permanently store your auth details in your .Rprofile by putting in
entries like this:

options(cloudant = list(user='name', pwd='pwd'))

Though if you don't want to store them permanently, you can use the cushion()
function instead. See examples below on how to do this. Though using cusion()
only stores them for the current session.

\code{profile()} Looks first in the local environment SofaAuthCache, and if finds nothing,
looks in your \code{.Rprofile} file.
}
\examples{
\donttest{
cushion('cloudant', user='name', pwd='pwd', type="cloudant")
cushion('iriscouch', user='name', pwd='pwd', type="iriscouch")
cushion('julies_iris', user='name', pwd='pwd', type="iriscouch")

cushion(localhost = list(type="localhost", port=5984))
profiles()
}
}

