% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authentication}
\alias{authentication}
\alias{cushion}
\alias{sofa_profile}
\title{Put a cushion on the sofa, and get cushion info}
\usage{
cushion(...)

sofa_profile()
}
\arguments{
\item{...}{Enter named sets of username and password.}
}
\description{
That is, set up config for remote CouchDB databases, or get auth info
}
\details{
Setup for authentication:
For localhost you don't need to authenticate, but of course you may have set
up a username and password in which case use 'localhost'.

Others supported are 'cloudant' and 'iriscouch'.

You can use other named username/password sets too.

You can permanently store your auth details in your .Rprofile by putting in
entries like this:

options(sofa_cloudant = c("username","yourcoolpassword"))

Though if you don't want to store them permanently, you can use the cushion()
function instead. See examples below on how to do this. Though using cusion()
only stores them for the current session.

\code{sofa_profile()} Looks first in the local environment SofaAuthCache, and if finds nothing,
looks in your \code{.Rprofile} file.
}
\examples{
\donttest{
cushion(sofa_cloudant=c('name','pwd'), sofa_iriscouch=c('name','pwd'))
sofa_profile()
}
}

