% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createdb}
\alias{createdb}
\title{Create a database.}
\usage{
createdb(dbname, endpoint = "localhost", port = 5984, username = NULL,
  pwd = NULL, delifexists = FALSE, ...)
}
\arguments{
\item{dbname}{Database name}

\item{endpoint}{One of localhost, cloudant, or iriscouch. For local work
use the default localhost. For cloudant or iriscouch you will need accounts
with those database services.}

\item{port}{The port to use. Only applicable if using endpoint="localhost".}

\item{username}{Your cloudant or iriscouch username}

\item{pwd}{Your cloudant or iriscouch password}

\item{delifexists}{If TRUE, delete any database of the same name before creating it.
This is useful for testing. Default is FALSE.}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Create a database.
}
\examples{
\donttest{
createdb(dbname='leothelion')
listdbs() # see if its there now

## or setting username and password in cushion() call
cushion(cloudant_name='name', cloudant_pwd='pwd')
createdb(dbname="mustache", "cloudant")

## iriscouch
cushion(iriscouch_name='name', iriscouch_pwd='pwd')
createdb(dbname="mustache", "iriscouch")

# different login credentials than normal, just pass in to function call
createdb(dbname='stuff', "cloudant", username='<user>', pwd='<pwd>')
}
}

