% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/databases.r
\name{databases}
\alias{databases}
\title{Work with databases in your CouchDB's.}
\description{
Work with databases in your CouchDB's.
}
\details{
There are the following functions for working with databases:
\itemize{
 \item \code{\link{db_create}} - Create a database
 \item \code{\link{db_delete}} - Delete a database
 \item \code{\link{db_info}} - Get info for a database
 \item \code{\link{db_list}} - List databases
 \item \code{\link{db_replicate}} - Replicate a database from one couch to another
}
}

