% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db_create.r
\name{db_create}
\alias{db_create}
\title{Create a database.}
\usage{
db_create(cushion = "localhost", dbname, delifexists = FALSE, as = "list",
  ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name}

\item{delifexists}{If TRUE, delete any database of the same name before creating it.
This is useful for testing. Default is FALSE.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Create a database.
}
\examples{
\donttest{
db_create(dbname='leothelion')
db_list() # see if its there now

## or setting username and password in cushion() call
db_create("cloudant", "mustache")

## iriscouch
db_create("iriscouch", "beard")
}
}

