% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db_delete.r
\name{db_delete}
\alias{db_delete}
\title{Delete a database.}
\usage{
db_delete(cushion = "localhost", dbname, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a database.
}
\examples{
\donttest{
# local databasees
## create database first, then delete
db_create(dbname='newdb')
db_delete(dbname='newdb')

# cloudant
db_create("cloudant", "stuffthings")
listdbs("cloudant")
db_delete("cloudant", "stuffthings")

## with curl info while doing request
library('httr')
db_create('newdb')
db_delete('newdb', config=verbose())
}
}

