% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.r
\name{db_info}
\alias{db_info}
\title{List database info.}
\usage{
db_info(cushion, dbname, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{Database name}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
List database info.
}
\examples{
\dontrun{
(x <- Cushion$new())

db_info(x, dbname="sofadb")
db_info(x, dbname="sofadb", as='json')
}
}

