% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db_list.r
\name{db_list}
\alias{db_list}
\title{List all databases.}
\usage{
db_list(cushion = "localhost", simplify = TRUE, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{simplify}{(logical) Simplify to character vector, ignored if \code{as="json"}}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
List all databases.
}
\examples{
\donttest{
# local databasees
db_list()
db_list(as='json')
db_list("cloudant")
db_list("iriscouch")
db_list("oceancouch")
}
}

