% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbinfo}
\alias{dbinfo}
\title{List database info.}
\usage{
dbinfo(dbname, endpoint = "localhost", port = 5984, username = NULL,
  pwd = NULL, ...)
}
\arguments{
\item{dbname}{Database name}

\item{endpoint}{One of localhost, cloudant, or iriscouch. For local work
use the default localhost. For cloudant or iriscouch you will need accounts
with those database services.}

\item{port}{The port to use. Only applicable if using endpoint="localhost".}

\item{username}{Your cloudant or iriscouch username}

\item{pwd}{Your cloudant or iriscouch password}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
List database info.
}
\examples{
\donttest{
# local databasees
dbinfo(dbname="sofadb")

# databasees on cloudant
## passing username and password in function call
dbinfo("cloudant", username='yourusername', pwd='yourpassword')

## or setting username and password in options() call
cushion(cloudant_username='name', cloudant_pwd='pwd')
dbinfo("gaugesdb_ro", "cloudant")
}
}

