% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deldoc}
\alias{deldoc}
\title{Delete a document in a database.}
\usage{
deldoc(endpoint = "http://127.0.0.1", port = 5984, dbname, docid)
}
\arguments{
\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}
}
\description{
Delete a document in a database.
}
\examples{
\donttest{
doc3 <- "<top><a/><b/><c><d/><e>bob</e></c></top>"
writedoc(dbname="sofadb", doc=doc3, docid="somexml")
deldoc(dbname="sofadb", docid="somexml")

# wrong docid name
writedoc(dbname="sofadb", doc=doc3, docid="somexml")
deldoc(dbname="sofadb", docid="wrongname")
}
}

