% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deletedb}
\alias{deletedb}
\title{Delete a database.}
\usage{
deletedb(cushion = "localhost", dbname, ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a database.
}
\examples{
\donttest{
# local databasees
## create database first, then delete
createdb('newdb')
deletedb('newdb')

# cloudant
createdb("cloudant", "stuffthings")
listdbs("cloudant")
deletedb("cloudant", "stuffthings")

## with curl info while doing request
library('httr')
createdb('newdb')
deletedb('newdb', config=verbose())
}
}

