% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doc_create}
\alias{doc_create}
\title{Write documents to a database.}
\usage{
doc_create(cushion = "localhost", dbname, doc, docid = NULL,
  apicall = FALSE, baseurl, queryargs, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name3}

\item{doc}{Document content}

\item{docid}{Document ID}

\item{apicall}{If TRUE, write with json format e.g.:
{
  "baseurl" : "http://alm.plos.org/api/v3/articles",
  "yourqueryargs" : "doi=10.1371/journal.pone.0060590",
  "response": "response_from_the_api"
}}

\item{baseurl}{Base url for the web API call}

\item{queryargs}{Web API query arguments to pass in to json with document}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Write documents to a database.
}
\examples{
\donttest{
# write a document WITH a name (uses PUT)
doc1 <- '{"name":"drink","beer":"IPA"}'
doc_create(dbname="sofadb", doc=doc1, docid="abeer")
doc_create(dbname="sofadb", doc=doc1, docid="morebeer", as='json')
doc_get(dbname = "sofadb", docid = "abeer")

# write a json document WITHOUT a name (uses POST)
doc2 <- '{"name":"food","icecream":"rocky road"}'
doc_create(dbname="sofadb", doc=doc2)

doc3 <- '{"planet":"mars","size":"smallish"}'
doc_create(dbname="sofadb", doc=doc3)

# write an xml document WITH a name (uses PUT). xml is written as xml in
# couchdb, just wrapped in json, when you get it out it will be as xml
doc4 <- "<top><a/><b/><c><d/><e>bob</e></c></top>"
doc_create(dbname="sofadb", doc=doc4, docid="somexml")
doc_get(dbname = "sofadb", docid = "somexml")

# write a document using web api storage format
doc <- '{"downloads":10,"pageviews":5000,"tweets":300}'
doc_create(dbname="sofadb", doc=doc, docid="asdfg", apicall=TRUE, baseurl="http://things...",
   queryargs="some args")
doc_get(dbname = "sofadb", docid = "asdfg")

# in iriscouch
doc_create("iriscouch", dbname='helloworld', doc='{"things":"stuff"}', docid="ggg")
doc_get("iriscouch", dbname='helloworld', docid="ggg")
doc_delete("iriscouch", dbname='helloworld', docid="ggg")
}
}

