% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.r
\name{documents}
\alias{documents}
\title{Work with documents in your CouchDB's.}
\description{
Work with documents in your CouchDB's.
}
\details{
If you are writing a complicated javascript function, better to do
that in the Futon CouchDB interface or otherwise.

There are the following functions for working with documents:
\itemize{
\item \code{doc_create} - Create a document, with or without an ID
\item \code{doc_update} - Update a document
\item \code{doc_get} - Get a document
\item \code{doc_delete} - Delete a document
\item \code{doc_head} - Get headers for a document
\item \code{doc_attach_create} - Attach something to a document
\item \code{doc_attach_info} - Get info on an attachment
\item \code{doc_attach_get} - Fetch an attachment
\item \code{doc_attach_delete} - Delete an attachment
\item \code{db_alldocs} - Get all documents
\item \code{db_revisions} - Get revisions for a document
}
}
