% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documents.r
\name{documents}
\alias{documents}
\title{Work with documents in your CouchDB's.}
\description{
Work with documents in your CouchDB's.
}
\details{
If you are writing a complicated javascript function, better to do
that in the Futon CouchDB interface or otherwise.

There are the following functions for working with documents:
\itemize{
 \item \code{\link{doc_create}} - Create a document, with or without an ID
 \item \code{\link{doc_update}} - Update a document
 \item \code{\link{doc_get}} - Get a document
 \item \code{\link{doc_delete}} - Delete a document
 \item \code{\link{doc_head}} - Get headers for a document
 \item \code{\link{attach_create}} - Attach something to a document
 \item \code{\link{alldocs}} - Get all documents
 \item \code{\link{revisions}} - Get revisions for a document
}
}

