% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getdoc}
\alias{getdoc}
\title{Get a document from a database.}
\usage{
getdoc(endpoint = "http://127.0.0.1", port = 5984, dbname, docid, ...)
}
\arguments{
\item{endpoint}{the endpoint, defaults to localhost (http://127.0.0.1)}

\item{port}{port to connect to, defaults to 5984}

\item{dbname}{Database name}

\item{docid}{Document ID}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get a document from a database.
}
\examples{
\donttest{
getdoc(dbname="sofadb", docid="beer")
}
}

