% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getdoc}
\alias{getdoc}
\title{Get a document from a database.}
\usage{
getdoc(cushion = "localhost", dbname, docid, ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name}

\item{docid}{Document ID}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get a document from a database.
}
\examples{
\donttest{
getdoc(dbname="sofadb", docid="beer")
getdoc("cloudant", dbname='gaugesdb_ro', docid='017ba8075b92656bbca20b8ab6fdb21d')
getdoc("iriscouch", dbname='helloworld', docid="0c0858b75a81c464a74119ca24000543")
}
}

