% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\title{membership}
\usage{
membership(cushion, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
membership
}
\examples{
\dontrun{
# Create a CouchDB connection client
(x <- Cushion$new())

membership(x)
membership(x, as = 'json')
}
}

