% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ping}
\alias{ping}
\title{Ping a couchdb server.}
\usage{
ping(cushion = "localhost", port = 5984, username = NULL, pwd = NULL,
  ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{port}{port to connect to, defaults to 5984}

\item{username}{Your cloudant or iriscouch username}

\item{pwd}{Your cloudant or iriscouch password}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Ping a couchdb server.
}
\examples{
\donttest{
ping()
ping("cloudant")
ping("iriscouch")
}
}

