% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ping.r
\name{ping}
\alias{ping}
\title{Ping a couchdb server.}
\usage{
ping(cushion = "localhost", as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Ping a couchdb server.
}
\examples{
\donttest{
ping()
ping("cloudant")
ping("iriscouch")
ping("oceancouch")
}
}

