% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.r
\name{revisions}
\alias{revisions}
\title{Get document revisions.}
\usage{
revisions(cushion, dbname, docid, simplify = TRUE, as = "list", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{Database name}

\item{docid}{Document ID}

\item{simplify}{(logical) Simplify to character vector of revision ids.
If \code{FALSE}, gives back availabilit info too. Default: \code{TRUE}}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get document revisions.
}
\examples{
\dontrun{
(x <- Cushion$new())

if ("sofa" \%in\% db_list(x)) {
 db_delete(x, dbname = "sofadb")
}
db_create(x, dbname = "sofadb")

doc1 <- '{"name": "drink", "beer": "IPA", "score": 5}'
doc_create(x, dbname="sofadb", doc1, docid="abeer")
doc_create(x, dbname="sofadb", doc1, docid="morebeer", as='json')

revisions(x, dbname="sofadb", docid="abeer")
revisions(x, dbname="sofadb", docid="abeer", simplify=FALSE)
revisions(x, dbname="sofadb", docid="abeer", as='json')
revisions(x, dbname="sofadb", docid="abeer", simplify=FALSE, as='json')
}
}

