% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sf_head}
\alias{sf_head}
\title{Get header info for a document.}
\usage{
sf_head(endpoint = "localhost", port = 5984, dbname, docid,
  username = NULL, pwd = NULL, ...)
}
\arguments{
\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}
}
\description{
Get header info for a document.
}
\examples{
\donttest{
sf_head(dbname="sofadb", docid="beer")
}
}

