% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sf_head}
\alias{sf_head}
\title{Get header info for a document.}
\usage{
sf_head(cushion = "localhost", dbname, docid, as = "list", ...)
}
\arguments{
\item{cushion}{A cushion name}

\item{dbname}{Database name. (charcter)}

\item{docid}{Document ID (character)}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get header info for a document.
}
\examples{
\donttest{
sf_head(dbname="sofadb", docid="a_beer")
sf_head(dbname="sofadb", docid="a_beer", as='json')
sf_head("cloudant", dbname="animaldb", docid="badger")
sf_head("iriscouch", dbname="helloworld", docid="ggg")
}
}

