% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sofa-package.r
\docType{package}
\name{sofa-package}
\alias{sofa}
\alias{sofa-package}
\title{R client for CouchDB}
\description{
R client for CouchDB.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

