% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.r
\name{views}
\alias{view_del}
\alias{view_get}
\alias{view_put}
\alias{view_search}
\alias{views}
\title{Work with design documents (i.e., views) in your CouchDB's.}
\usage{
view_put(cushion, dbname, design_name, fxnname = "foo", key = "null",
  value = "doc", as = "json", ...)

view_del(cushion, dbname, design_name, as = "json", ...)

view_get(cushion, dbname, design_name, as = "json", ...)

view_search(cushion, dbname, design_name, query = NULL, as = "json", ...)
}
\arguments{
\item{cushion}{A \code{Cushion} object. Required.}

\item{dbname}{(character) Database name}

\item{design_name}{(character) Design document name}

\item{fxnname}{(character) A function name}

\item{key}{(character) Key}

\item{value}{(character) Value}

\item{as}{(character) One of list (default) or json}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}

\item{query}{(character) Query terms}
}
\description{
Work with design documents (i.e., views) in your CouchDB's.
}
\details{
If you are writing a complicated javascript function, better to do
that in the Futon CouchDB interface or otherwise.
}
\examples{
\dontrun{
# Create a view
db_create(dbname = "alm_couchdb")

view_put(dbname='alm_couchdb', design_name='almview1')
view_put(dbname='alm_couchdb', design_name='almview2', value="doc.baseurl")
view_put(dbname='alm_couchdb', design_name='almview5', value="[doc.baseurl,doc.queryargs]")

# Delete a view
view_del(dbname='alm_couchdb', design_name='almview1')

# Get info on a design document (i.e., a view)
# view_get(dbname='alm_couchdb', design_name='almview1')

# Search using a view
# view_search(dbname='alm_couchdb', design_name='almview1', query="XXX")
}
}

