% Generated by roxygen2 (4.0.2): do not edit by hand
\name{views}
\alias{view_del}
\alias{view_get}
\alias{view_put}
\alias{view_search}
\alias{views}
\title{Work with design documents (i.e., views) in your CouchDB's.}
\usage{
view_put(dbname, design_name, fxnname = "foo", key = "null",
  value = "doc", endpoint = "http://127.0.0.1", port = 5984)

view_del(dbname, design_name, endpoint = "http://127.0.0.1", port = 5984)

view_get(dbname, design_name, endpoint = "http://127.0.0.1", port = 5984)

view_search(dbname, design_name, query = NULL,
  endpoint = "http://127.0.0.1", port = 5984)
}
\arguments{
\item{dbname}{Database name}

\item{design_name}{Design document name}

\item{fxnname}{A function name}

\item{key}{Key}

\item{value}{Value}

\item{endpoint}{the endpoint, defaults to localhost (http://127.0.0.1)}

\item{port}{port to connect to, defaults to 5984}

\item{query}{Query terms}
}
\description{
Work with design documents (i.e., views) in your CouchDB's.
}
\details{
If you are writing a complicated javascript function, better to do
   that in the Futon CouchDB interface or otherwise.
}
\examples{
\donttest{
# Create a view
view_put(dbname='alm_couchdb', design_name='almview1')
view_put(dbname='alm_couchdb', design_name='almview2', value="doc.baseurl")
view_put(dbname='alm_couchdb', design_name='almview5', value="[doc.baseurl,doc.queryargs]")

# Delete a view
view_del(dbname='alm_couchdb', design_name='almview1')

# Get info on a design document (i.e., a view)
view_get(dbname='alm_couchdb', design_name='almview1')

# Search using a view
view_search(dbname='alm_couchdb', design_name='almview1', query="XXX")
}
}

